# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KINECTRA API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SessionInputAnalysisType = typeof SessionInputAnalysisType[keyof typeof SessionInputAnalysisType];


export const SessionInputAnalysisType = {
  bowling: 'bowling',
  batting: 'batting',
} as const;

export type SessionInputSkillLevel = typeof SessionInputSkillLevel[keyof typeof SessionInputSkillLevel];


export const SessionInputSkillLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
  professional: 'professional',
} as const;

export type SessionInputDominantHand = typeof SessionInputDominantHand[keyof typeof SessionInputDominantHand];


export const SessionInputDominantHand = {
  right: 'right',
  left: 'left',
} as const;

export interface SessionInput {
  athleteName: string;
  analysisType: SessionInputAnalysisType;
  skillLevel: SessionInputSkillLevel;
  dominantHand: SessionInputDominantHand;
}

export type SessionAnalysisType = typeof SessionAnalysisType[keyof typeof SessionAnalysisType];


export const SessionAnalysisType = {
  bowling: 'bowling',
  batting: 'batting',
} as const;

export type SessionSkillLevel = typeof SessionSkillLevel[keyof typeof SessionSkillLevel];


export const SessionSkillLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
  professional: 'professional',
} as const;

export type SessionDominantHand = typeof SessionDominantHand[keyof typeof SessionDominantHand];


export const SessionDominantHand = {
  right: 'right',
  left: 'left',
} as const;

export type SessionStatus = typeof SessionStatus[keyof typeof SessionStatus];


export const SessionStatus = {
  active: 'active',
  completed: 'completed',
} as const;

export interface Session {
  id: string;
  athleteName: string;
  analysisType: SessionAnalysisType;
  skillLevel: SessionSkillLevel;
  dominantHand: SessionDominantHand;
  createdAt: string;
  status: SessionStatus;
}

export type SnapshotMetrics = {
  elbowAngle?: number;
  spineTilt?: number;
  kneeAngle?: number;
  shoulderAlignment?: number;
};

export interface Snapshot {
  src: string;
  label: string;
  time: string;
  category?: string;
  metrics?: SnapshotMetrics;
}

export interface SessionEndInput {
  frameCount: number;
  avgPostureScore: number;
  avgAlignmentScore: number;
  avgStabilityScore: number;
  avgEfficiencyScore: number;
  overallScore: number;
  warnings: string[];
  snapshots?: Snapshot[];
}

export type SessionSummaryAnalysisType = typeof SessionSummaryAnalysisType[keyof typeof SessionSummaryAnalysisType];


export const SessionSummaryAnalysisType = {
  bowling: 'bowling',
  batting: 'batting',
} as const;

export type SessionSummarySkillLevel = typeof SessionSummarySkillLevel[keyof typeof SessionSummarySkillLevel];


export const SessionSummarySkillLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
  professional: 'professional',
} as const;

export type SessionSummaryDominantHand = typeof SessionSummaryDominantHand[keyof typeof SessionSummaryDominantHand];


export const SessionSummaryDominantHand = {
  right: 'right',
  left: 'left',
} as const;

export type SessionSummaryStatus = typeof SessionSummaryStatus[keyof typeof SessionSummaryStatus];


export const SessionSummaryStatus = {
  active: 'active',
  completed: 'completed',
} as const;

export interface SessionSummary {
  id: string;
  athleteName: string;
  analysisType: SessionSummaryAnalysisType;
  skillLevel: SessionSummarySkillLevel;
  dominantHand: SessionSummaryDominantHand;
  createdAt: string;
  status: SessionSummaryStatus;
  frameCount: number;
  avgPostureScore: number;
  avgAlignmentScore: number;
  avgStabilityScore: number;
  avgEfficiencyScore: number;
  overallScore: number;
  strengths: string[];
  improvements: string[];
  recommendations: string[];
  warnings: string[];
  snapshots?: Snapshot[];
}

export interface ErrorResponse {
  error: string;
}