# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KINECTRA API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: session
    description: Analysis session operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /session/start:
    post:
      operationId: startSession
      tags: [session]
      summary: Start an analysis session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionInput"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"

  /session/{sessionId}/end:
    post:
      operationId: endSession
      tags: [session]
      summary: End an analysis session and get summary
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionEndInput"
      responses:
        "200":
          description: Session ended with summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionSummary"

  /session/{sessionId}:
    get:
      operationId: getSession
      tags: [session]
      summary: Get session results
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionSummary"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /session:
    get:
      operationId: listSessions
      tags: [session]
      summary: List recent sessions
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    SessionInput:
      type: object
      required:
        - athleteName
        - analysisType
        - skillLevel
        - dominantHand
      properties:
        athleteName:
          type: string
        analysisType:
          type: string
          enum: [bowling, batting]
        skillLevel:
          type: string
          enum: [beginner, intermediate, advanced, professional]
        dominantHand:
          type: string
          enum: [right, left]

    Session:
      type: object
      required:
        - id
        - athleteName
        - analysisType
        - skillLevel
        - dominantHand
        - createdAt
        - status
      properties:
        id:
          type: string
        athleteName:
          type: string
        analysisType:
          type: string
          enum: [bowling, batting]
        skillLevel:
          type: string
          enum: [beginner, intermediate, advanced, professional]
        dominantHand:
          type: string
          enum: [right, left]
        createdAt:
          type: string
        status:
          type: string
          enum: [active, completed]

    SessionEndInput:
      type: object
      required:
        - frameCount
        - avgPostureScore
        - avgAlignmentScore
        - avgStabilityScore
        - avgEfficiencyScore
        - overallScore
        - warnings
      properties:
        frameCount:
          type: integer
        avgPostureScore:
          type: number
        avgAlignmentScore:
          type: number
        avgStabilityScore:
          type: number
        avgEfficiencyScore:
          type: number
        overallScore:
          type: number
        warnings:
          type: array
          items:
            type: string
        snapshots:
          type: array
          items:
            $ref: "#/components/schemas/Snapshot"

    SessionSummary:
      type: object
      required:
        - id
        - athleteName
        - analysisType
        - skillLevel
        - dominantHand
        - createdAt
        - status
        - frameCount
        - avgPostureScore
        - avgAlignmentScore
        - avgStabilityScore
        - avgEfficiencyScore
        - overallScore
        - strengths
        - improvements
        - recommendations
        - warnings
      properties:
        id:
          type: string
        athleteName:
          type: string
        analysisType:
          type: string
          enum: [bowling, batting]
        skillLevel:
          type: string
          enum: [beginner, intermediate, advanced, professional]
        dominantHand:
          type: string
          enum: [right, left]
        createdAt:
          type: string
        status:
          type: string
          enum: [active, completed]
        frameCount:
          type: integer
        avgPostureScore:
          type: number
        avgAlignmentScore:
          type: number
        avgStabilityScore:
          type: number
        avgEfficiencyScore:
          type: number
        overallScore:
          type: number
        strengths:
          type: array
          items:
            type: string
        improvements:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        snapshots:
          type: array
          items:
            $ref: "#/components/schemas/Snapshot"

    Snapshot:
      type: object
      required:
        - src
        - label
        - time
      properties:
        src:
          type: string
        label:
          type: string
        time:
          type: string
        category:
          type: string
        metrics:
          type: object
          properties:
            elbowAngle:
              type: number
            spineTilt:
              type: number
            kneeAngle:
              type: number
            shoulderAlignment:
              type: number

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string